修改配置文件

+++ cli-generator/src/main/java/com/dmj/cli/domain/BusinessCooperation.java
@@ -29,10 +29,10 @@ public class BusinessCooperation extends BaseEntity implements Serializable {
     private static final long serialVersionUID=1L;
     
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @TableField("type_id")
-    private Integer typeId;
+    private Long typeId;
 
     @ApiModelProperty(value = "标题")
     @TableField("title")
